Добавление шапки, исправление архитектуры
Исправил архитектуру кода - вынес шапку документации в файл со всеми шапками экранов, а так же добавил в нем шапку экрана "история"
