+++ habits/lib/pages/history.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:habits/elements/appBars.dart';
 
 class HistoryPage extends StatelessWidget {
   const HistoryPage({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
+    return Scaffold(
+      appBar: histAppBar(),
+      body: const Center(
         child: Text(
           'History',
           style: TextStyle(fontSize: 30),
